pkg/asset/installconfig: Push platform-specific logic into subdirectories
To keep the platform-specific code separate.

The only thing that wasn't perfectly clean about this separation is
that now the AWS and OpenStack packages have their own defaultVPCCIDR.
There's no reason that they need to agree on that CIDR, though, so I
haven't bothered setting up a common-default package for them to share
a single constant.
